# Improved performance and memory usage