--- a/Mipa.Benchmark/Program.cs
+++ b/Mipa.Benchmark/Program.cs
@@ -1,0 +1,6 @@
+﻿// See https://aka.ms/new-console-template for more information
+
+using BenchmarkDotNet.Running;
+using Mipa.Benchmark;
+
+BenchmarkRunner.Run<ParseBenchmark>();