--- a/Mipa.Tests/Mipa.Tests.csproj
+++ b/Mipa.Tests/Mipa.Tests.csproj
@@ -13,7 +13,6 @@
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.6.0"/>
         <PackageReference Include="NUnit" Version="3.13.3"/>
         <PackageReference Include="NUnit3TestAdapter" Version="4.2.1"/>
-        <PackageReference Include="NUnit.Analyzers" Version="3.6.1"/>
         <PackageReference Include="coverlet.collector" Version="6.0.0"/>
     </ItemGroup>