--- a/Mipa.Tests/ParserTests.cs
+++ b/Mipa.Tests/ParserTests.cs
@@ -1,6 +1,6 @@
 namespace Mipa.Tests;
 
-public record UnparsableType : IParsable<UnparsableType>
+public record UnparsableType : IParsable<UnparsableType>, ISpanParsable<UnparsableType>
 {
     public static UnparsableType Parse(string s, IFormatProvider? provider)
     {
@@ -8,6 +8,17 @@ public record UnparsableType : IParsable<UnparsableType>
     }
 
     public static bool TryParse(string value, IFormatProvider? formatProvider, out UnparsableType result)
+    {
+        result = new UnparsableType();
+        return false;
+    }
+
+    public static UnparsableType Parse(ReadOnlySpan<char> s, IFormatProvider? provider)
+    {
+        throw new Exception();
+    }
+
+    public static bool TryParse(ReadOnlySpan<char> s, IFormatProvider? provider, out UnparsableType result)
     {
         result = new UnparsableType();
         return false;
@@ -23,7 +34,7 @@ public class Tests
         var result = parser.Parse("");
         
         Assert.That(result.Content, Is.EqualTo(""));
-        Assert.That(result.Arguments, Is.Empty);
+        Assert.That(result.Arguments, Has.Length.Zero);
     }
     
     [Test]
@@ -33,7 +44,7 @@ public class Tests
         var result = parser.Parse("hello world");
         
         Assert.That(result.Content, Is.EqualTo("hello world"));
-        Assert.That(result.Arguments, Is.Empty);
+        Assert.That(result.Arguments, Has.Length.Zero);
     }
     
     [Test]
@@ -43,10 +54,21 @@ public class Tests
         var result = parser.Parse("hello world key:value");
         
         Assert.That(result.Content, Is.EqualTo("hello world"));
-        Assert.That(result.Arguments, Has.Count.EqualTo(1));
-        Assert.That(result.GetArgument("key"), Is.EqualTo("value"));
+        Assert.That(result.Arguments, Has.Length.EqualTo(1));
+        Assert.That(result.GetArgument("key").Span is "value");
     }
-    
+
+    [Test]
+    public void SingleArgumentBetweenContent()
+    {
+        var parser = new MipaParser();
+        var result = parser.Parse("hello key:value world");
+
+        Assert.That(result.Content, Is.EqualTo("hello world"));
+        Assert.That(result.Arguments, Has.Length.EqualTo(1));
+        Assert.That(result.GetArgument("key").Span is "value");
+    }
+
     [Test]
     public void MultipleArguments()
     {
@@ -54,9 +76,9 @@ public class Tests
         var result = parser.Parse("key2:value2 hello world key:value ");
         
         Assert.That(result.Content, Is.EqualTo("hello world"));
-        Assert.That(result.Arguments, Has.Count.EqualTo(2));
-        Assert.That(result.GetArgument("key"), Is.EqualTo("value"));
-        Assert.That(result.GetArgument("key2"), Is.EqualTo("value2"));
+        Assert.That(result.Arguments, Has.Length.EqualTo(2));
+        Assert.That(result.GetArgument("key").Span is "value");
+        Assert.That(result.GetArgument("key2").Span is "value2");
     }
     
     [Test]
@@ -66,9 +88,9 @@ public class Tests
         var result = parser.Parse("key2:\"this is value2\" hello world key:\"value\" ");
         
         Assert.That(result.Content, Is.EqualTo("hello world"));
-        Assert.That(result.Arguments, Has.Count.EqualTo(2));
-        Assert.That(result.GetArgument("key"), Is.EqualTo("value"));
-        Assert.That(result.GetArgument("key2"), Is.EqualTo("this is value2"));
+        Assert.That(result.Arguments, Has.Length.EqualTo(2));
+        Assert.That(result.GetArgument("key").Span is "value");
+        Assert.That(result.GetArgument("key2").Span is "this is value2");
     }
 
     [Test]
@@ -78,7 +100,7 @@ public class Tests
         var result = parser.Parse("date:08/08/2023 hello world key:23");
 
         Assert.That(result.Content, Is.EqualTo("hello world"));
-        Assert.That(result.Arguments, Has.Count.EqualTo(2));
+        Assert.That(result.Arguments, Has.Length.EqualTo(2));
         Assert.That(result.GetArgument<DateTime>("date"), Is.TypeOf<DateTime>());
         Assert.That(result.GetArgument<DateTime>("date"), Is.EqualTo(new DateTime(2023, 8, 8)));
         Assert.That(result.GetArgument<int>("key"), Is.EqualTo(23));