--- a/Mipa/ParseResult.cs
+++ b/Mipa/ParseResult.cs
@@ -1,26 +1,39 @@
+using System.Collections.Immutable;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
 namespace Mipa;
 
-public record ParseResult
+public readonly struct ParseResult
 {
     /// <summary>
     /// Remainder of the content that is unparsed.
     /// </summary>
     public string Content { get; init; }
-    public IReadOnlyList<KeyValuePair<string, string>> Arguments { get; init; }
+    public ReadOnlyMemory<KeyValuePair<ReadOnlyMemory<char>, ReadOnlyMemory<char>>> Arguments { get; init; }
     
-    public string? GetArgument(string key)
+    public ReadOnlyMemory<char> GetArgument(string key)
     {
-        return Arguments.FirstOrDefault(x => x.Key == key).Value;
+        foreach (var argument in Arguments.Span)
+        {
+            if (argument.Key.Span.SequenceEqual(key))
+            {
+                return argument.Value;
+            }
+        }
+
+        return default;
     }
-    public T? GetArgument<T>(string key) where T : IParsable<T>
+
+    public T? GetArgument<T>(string key) where T : ISpanParsable<T>
     {
         var argument = GetArgument(key);
-        if (argument == null)
+        if (argument.IsEmpty)
         {
             return default;
         }
 
-        if (!T.TryParse(argument, null, out var result))
+        if (!T.TryParse(argument.Span, null, out var result))
         {
             return default;
         }