**Before**
| Method                            | Mean       | Error     | StdDev    | Gen0   | Allocated |
|---------------------------------- |-----------:|----------:|----------:|-------:|----------:|
| ParseEmpty                        |   5.373 ns | 0.0493 ns | 0.0461 ns | 0.0019 |      32 B |
| ParseHelloWorld                   | 160.102 ns | 1.7004 ns | 1.5074 ns | 0.0181 |     304 B |
| ParseSingleArgument               | 389.493 ns | 5.1651 ns | 4.8315 ns | 0.0644 |    1080 B |
| ParseSingleArgumentBetweenContent | 349.253 ns | 5.6797 ns | 5.3128 ns | 0.0615 |    1032 B |
| ParseMultipleArguments            | 522.896 ns | 6.9449 ns | 5.4221 ns | 0.1001 |    1680 B |
| ParseMultipleArgumentsWithQuotes  | 525.849 ns | 4.1642 ns | 3.8952 ns | 0.1078 |    1816 B |

**After**
| Method                           | Mean      | Error     | StdDev    | Gen0   | Allocated |
|--------------------------------- |----------:|----------:|----------:|-------:|----------:|
| ParseEmpty                       |  9.522 ns | 0.0763 ns | 0.0714 ns |      - |         - |
| ParseHelloWorld                  | 10.560 ns | 0.0563 ns | 0.0470 ns |      - |         - |
| ParseSingleArgument              | 46.161 ns | 0.5124 ns | 0.4001 ns | 0.0062 |     104 B |
| ParseMultipleArguments           | 72.213 ns | 0.8378 ns | 0.7837 ns | 0.0081 |     136 B |
| ParseMultipleArgumentsWithQuotes | 80.293 ns | 1.6282 ns | 1.6720 ns | 0.0081 |     136 B |

A lot of breaking changes have been made.
Also `"hello key:value world"` currently parses to the content `"helloworld"`, which is a bug. In the main branche it parses to `"hello  world"`.